requirements.txt
decorator==4.4.1
Django==3.1.1
djangorestframework==3.12.1
Pillow==7.0.0
pylint==2.4.4